Kill React warning by not passing setStack prop
Otherwise you get:

Warning: React does not recognize the `setStack` prop on a DOM element. If you intentionally want it to appear in the DOM as a custom attribute, spell it as lowercase `setstack` instead. If you accidentally passed it from a parent component, remove it from the DOM element.